requirements.txt
# Layer10 Memory Graph - Dependencies
# All free/open-source

# API & HTTP
requests>=2.31.0
python-dotenv>=1.0.0

# LLM (Groq free tier)
groq>=0.4.0

# Database
# sqlite3 is built-in

# NLP & Embeddings
sentence-transformers>=2.2.0
spacy>=3.7.0

# Graph
networkx>=3.2.0
pyvis>=0.3.2

# Visualization
streamlit>=1.29.0

# Data processing
numpy>=1.26.0
pandas>=2.1.0
tqdm>=4.66.0

# Utilities
python-dateutil>=2.8.0
